# src/parking/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parking API",
        "description": "API for parking",
        "version": "1.0.0"
    },
    "paths": {
        "/api/parking-lots": {
            "get": {
                "tags": [
                    "Parking Lots"
                ],
                "summary": "List parking lots",
                "description": "Retrieve a list of all parking lots with their associated parking spots",
                "operationId": "listParkingLots",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ParkingLot"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/parking-lots/{id}": {
            "get": {
                "tags": [
                    "Parking Lots"
                ],
                "summary": "Get parking lot details",
                "description": "Retrieve the details of a specific parking lot, including its associated parking spots",
                "operationId": "showParkingLot",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the parking lot",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParkingLot"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parking lot ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid parking lot ID"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parking lot not found"
                    }
                }
            }
        },
        "/api/parking-lots/{id}/availability": {
            "get": {
                "tags": [
                    "Parking Lots"
                ],
                "summary": "Get parking lot availability",
                "description": "Retrieve the total spots and available spots for a specific parking lot",
                "operationId": "getAvailability",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the parking lot",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_spots": {
                                            "type": "integer"
                                        },
                                        "available_spots": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parking lot ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid parking lot ID"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parking lot not found"
                    }
                }
            }
        },
        "/api/parking-spots/{id}/park": {
            "post": {
                "summary": "Park a vehicle in a parking spot",
                "operationId": "2fa13683ea963121a56db008399ceda1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the parking spot",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vehicle_type",
                        "in": "query",
                        "description": "Type of the vehicle",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "parking_lot_id",
                        "in": "query",
                        "description": "ID of the parking lot",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle parked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vehicle parked successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parking spot ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid parking spot ID"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No available spot found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No available spot found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/parking-spots/{id}/unpark": {
            "post": {
                "summary": "Unpark a vehicle from a parking spot",
                "operationId": "c7bd9facc3a4a1cd0f28aae5728c566b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the parking spot",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle unparked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vehicle unparked successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parking spot ID or spot is already free",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": [
                                                "Invalid parking spot ID",
                                                "Spot is already free"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parking spot not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Parking spot not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ParkingLot": {
                "required": [
                    "id",
                    "name",
                    "address",
                    "city",
                    "state",
                    "zip",
                    "capacity"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Main Street Parking"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "city": {
                        "type": "string",
                        "example": "Anytown"
                    },
                    "state": {
                        "type": "string",
                        "example": "FL"
                    },
                    "zip": {
                        "type": "string",
                        "example": "12345"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
            "ParkingSpot": {
                "required": [
                    "id",
                    "occupied",
                    "parking_lot_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "occupied": {
                        "type": "string",
                        "example": "true"
                    },
                    "parking_lot_id": {
                        "type": "integer",
                        "example": "10"
                    }
                },
                "type": "object"
            },
            "Vehicle": {
                "required": [
                    "id",
                    "type",
                    "parking_spot_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "Car"
                    },
                    "parking_spot_id": {
                        "type": "integer",
                        "example": "10"
                    },
                    "plate_number": {
                        "type": "string",
                        "example": "10"
                    }
                },
                "type": "object"
            }
        }
    }
}